add mapping by method make router struct private rename package
